--- frontend/src/router/admin.js
@@ -1,9 +1,0 @@
-import slotRoute from "./slot";
-
-export const admin = [
-  {
-    path: "/admin",
-    name: "login",
-    children: [...slotRoute],
-  },
-];
