+++ frontend/src/views/admin/AboutView.vue
