+++ frontend/src/views/admin/SlotList.vue
