+++ frontend/src/views/admin/SpaceList.vue
